Fractal tree integration

See merge request cse1105/2018-2019/oopp-group-70/template!44

+++ src/main/java/controller/FoodController.java
@@ -26,7 +26,7 @@ public class FoodController {
         int vegmeals = model.getVegMealsCount();
         view.updateVegCounter(vegmeals);
 
-        float local = model.getLocalProduceCount();
+        double local = model.getLocalProduceCount();
         view.updateLocalProduceCounter(local);
 
     }
