+++ src/main/java/controller/PaperRecyclingController.java
@@ -24,7 +24,7 @@ public class PaperRecyclingController {
      */
     public void updateViewWithDate() {
 
-        float paperRecycling = model.getPaperRecyclingCount();
+        double paperRecycling = model.getPaperRecyclingCount();
         view.updatePaperRecyclingCounter(paperRecycling);
     }
 
@@ -34,9 +34,9 @@ public class PaperRecyclingController {
      * @param amountString - amount of the produced recycled paper
      */
     public void addPaperRecycling(String amountString) {
-        float kg = 0;
+        double kg = 0;
         try {
-            kg = Float.parseFloat(amountString);
+            kg = Double.parseDouble(amountString);
             if (kg < 0) {
                 view.displayStatus("Please enter the weight above 0 kg");
                 return;
