+++ src/main/java/model/objects/PaperRecyclingRequest.java
@@ -3,9 +3,9 @@ package model.objects;
 public class PaperRecyclingRequest {
 
     private String uid;
-    private float amount;
+    private double amount;
 
-    public PaperRecyclingRequest(String uid, float amount) {
+    public PaperRecyclingRequest(String uid, double amount) {
         this.uid = uid;
         this.amount = amount;
     }
@@ -14,7 +14,7 @@ public class PaperRecyclingRequest {
         return uid;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 }
